moved formatting to its own util
